Move endpoints for Django
Move existing endpoints from canonicalwebteam.yaml_redirects to
canonicalwebteam.yaml_redirects.django.
